fix: pf_putp function uses unsigned long type for pointer representation

+++ ft_puthp.c
@@ -6,7 +6,7 @@
 /*   By: vmistry <vmistry@student.42london.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/16 13:21:13 by vmistry           #+#    #+#             */
-/*   Updated: 2026/01/17 19:20:52 by vmistry          ###   ########.fr       */
+/*   Updated: 2026/01/17 19:26:47 by vmistry          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,11 +18,11 @@ void	pf_puth(unsigned long num, int *count, int upper)
 		pf_puth(num / 16, count, upper);
 	if (upper == 1)
 		pf_putc(ft_toupper(HEX[num % 16]), count);
-		else
+	else
 		pf_putc(HEX[num % 16], count);
 }
 
-void	pf_putp(unsigned long long p, int *count)
+void	pf_putp(unsigned long p, int *count)
 {
 	if (!p)
 		return (pf_puts("(nil)", count));
